NFC: Make stack heterogeneous
Stack is currently able to hold basic blocks as well as registers.
It's required for control flow instructions since they use
a continuation (basic block) from the stack to perform a jump.

The patch also extends std::variant<Ts...> providing operator==
to compare a variant with an object of a type constituting it.