Fix return optimization
The backend used return as an alias for NOP and didn't print it
to assembly. However the only correct omission of return could be done
in the exit BB of a function. The patch fixes return by introducing
proper opcode and printing RETURN_VOID and RETURN_I64. It also
introduces 2 optimizations:
- Remove POP s1 preceding RETARGS instruction.
- For exit basic block if there is no POP s1 to optimize away,
  remove RETARGS.