Fix of code generation for global address and external symbol operands
This commit fixes an issue with operands for global addresses and external symbols. In the existing implementation these addresses are added immediately before corresponding instruction using PUSHINT $symbol$ without tracking of pushed data in stack model. This leads to invalid stack state for instructions where global address is not last operand (for example, store requires address to be the first operand with second operand for value).

To fix this behaviour WebAssembly based lowering for global addresses and external symbols has been used:
- GLOBAL_ADDRESS_WRAPPER SDNode has been added;
- global address and external symbols have been lowered to a chain with GLOBAL_ADDRESS_WRAPPER and TargetGlobalAddress and TargetExternalSymbol nodes (so at the instruction level only GLOBAL_ADDRESS_WRAPPER can have operands with global address or external symbol which code generation is safe to add PUSHINT $symbol$ immediately before instruction);
- GLOBAL_ADDRESS_WRAPPER is lowered to fake instruction with opcode PUSH_GLOBAL_ADDRESS which is used as a marker to generate PUSHINT $symbol$ before it and is removed from instructions emitting process;
- because result of PUSH_GLOBAL_ADDRESS is a virtual register, it works perfectly ok with stack model.