Fix stack modeling for CFG
Introduce a pass for each basic block having 2 or more predecessors
excluding BBs that are loop headers inserts an empty basic block in between of
the processing BB and predecessor. Thus a newly created BB replaces the
original one for the processing BB and it becomes a successor of the
original predecessor.
Later TVM backend triggers stack cleaning logic in the beginning of each BB
(including newly created once) and removes all the registers that are
dead. It prevents cases when dead variables pass though an edge in CFG.
Consider a terminator br i1 %cond, label %bb1, label %bb2. Liveliness
analysis detect that after br both live variables of bb1 and bb2 has to
be kept in stack we pass to these basic block, since the compiler can't
know statically which branch will be chosen. It breaks the invariant that
only live registers are passed through CFG and ultimately leads to crash
on cfg1 test added to cfg.ll. The pass introduces the fix by removing
dead variable in the BBs inserted between the original predecessor and
a basic block.
The pass makes the CFG more complicated, and it should be fixed either
by a separate blocks removing / inlining redundant basic blocks or by
trying to exploit the following pattern in the pass itself:
PUSHCONT {bb1}
IFJMP
<stack cleanup since we know the right successor here>
PUSHCONT {bb2}
JMPX
Control-flow prepare pass breaks Codegen prepare invariants which needs
to be investigated. The commit disables the this pass by now.